Revert "refactor: better implementation of pixels (#4198)"
This reverts commit 40560b4916c91fa5eb13dae97c21c552f7bf85a1.
